move interleave serialization tests to kernel tests
