+++ tensorflow/python/data/experimental/kernel_tests/serialization/BUILD
@@ -86,26 +86,6 @@ tf_py_test(
 )
 
 tf_py_test(
-    name = "interleave_dataset_serialization_test",
-    size = "medium",
-    srcs = ["interleave_dataset_serialization_test.py"],
-    tags = [
-        "no_oss",
-        "no_pip",
-        "no_windows",
-    ],
-    deps = [
-        "//tensorflow/python:client_testlib",
-        "//tensorflow/python:sparse_ops",
-        "//tensorflow/python:sparse_tensor",
-        "//tensorflow/python/data/kernel_tests:checkpoint_test_base",
-        "//tensorflow/python/data/ops:dataset_ops",
-        "//third_party/py/numpy",
-        "@absl_py//absl/testing:parameterized",
-    ],
-)
-
-tf_py_test(
     name = "rebatch_dataset_serialization_test",
     size = "medium",
     srcs = ["rebatch_dataset_serialization_test.py"],
